Update composer.json
Hard coding the path to shortcode-button can cause problems if the parent project (for which wds-shortcodes is a dependency) also has shortcode-button as a dependency. In this case shortcode-button will be installed at the parent plugin vendor level and not at the wds-shortcodes level.

At least I think that is how it works.